day_24_youtube_gmail/requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated==0.0.2
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
asgiref==3.7.2
attrs==23.2.0
audioread==3.0.1
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.4
decorator==4.4.2
Deprecated==1.2.14
diffusers==0.26.3
diskcache==5.6.3
distro==1.9.0
Django==5.0.3
docker==7.0.0
faiss-cpu==1.8.0
fastapi==0.110.0
filelock==3.13.1
FLAML==2.1.1
flatbuffers==24.3.7
frozendict==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
googletrans==4.0.0rc1
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
html5lib==1.1
httpcore==1.0.4
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
image==1.5.33
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.11.0
importlib_resources==6.3.0
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.12
langchain-community==0.0.28
langchain-core==0.1.31
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langsmith==0.1.24
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
lxml==5.1.0
Markdown==3.6
markdown-it-py==3.0.0
markdownify==0.11.6
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
peewee==3.17.1
peft==0.9.0
pillow==10.2.0
platformdirs==4.2.0
pooch==1.8.1
posthog==3.5.0
praw==7.7.1
prawcore==2.4.0
proglog==0.1.10
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.19
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
pyparsing==3.1.2
PyPika==0.48.9
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.28
sqlparse==0.4.4
starlette==0.36.3
streamlit==1.32.1
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
transformers==4.38.2
typer==0.9.0
types-requests==2.31.0.20240311
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
yarl==1.9.4
yfinance==0.2.37
youtube-transcript-api==0.6.2
yt-dlp==2024.3.10
zipp==3.17.0
